# client-code-gen/app-access-control-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Forest Access Management - FAM - API",
    "description": "\nForest Access Management API used by the Forest Access Management application\nto Define who has access to what apps, and what roles they will operate under\n once access is granted.\n",
    "contact": {
      "name": "Team Heartwood",
      "url": "https://apps.nrs.gov.bc.ca/int/confluence/display/FSAST1/Team+Heartwood",
      "email": "SIBIFSAF@victoria1.gov.bc.ca"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/user-role-assignment": {
      "post": {
        "tags": [
          "FAM User Role Assignment"
        ],
        "summary": "Grant multiple users access to an application's role.",
        "description": "Granting IDIR/BCeID users access to an application's role, supporting expiry dates for role assignments.",
        "operationId": "create_user_role_assignment_many",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FamUserRoleAssignmentCreateSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FamUserRoleAssignmentRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "7qtdtska897dt6kandveusvd88": []
          }
        ]
      }
    },
    "/user-role-assignment/{user_role_xref_id}": {
      "delete": {
        "tags": [
          "FAM User Role Assignment"
        ],
        "summary": "Delete User Role Assignment",
        "description": "Remove a specific application's role from user's access.",
        "operationId": "delete_user_role_assignment",
        "security": [
          {
            "7qtdtska897dt6kandveusvd88": []
          }
        ],
        "parameters": [
          {
            "name": "user_role_xref_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Role Xref Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/districts": {
      "get": {
        "tags": [
          "FAM Districts"
        ],
        "summary": "Get Districts",
        "description": "List the BC natural resource districts available for scoping roles that\nhave role_type_district set.\n\nDistricts are a fixed reference set held in the District enum, not database\nrecords, so this returns the full list and expects the caller to filter.",
        "operationId": "get_districts",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FamDistrictSchema"
                  },
                  "type": "array",
                  "title": "Response Get Districts"
                }
              }
            }
          }
        },
        "security": [
          {
            "7qtdtska897dt6kandveusvd88": []
          }
        ]
      }
    },
...
    "/user-terms-conditions": {
      "post": {
        "tags": [
          "FAM User Terms and Conditions"
        ],
        "summary": "Create User Terms And Conditions",
        "description": "Create a record for terms and conditions acceptance. \n\nIf no version is provided, we store the 1st version of the terms and conditions.",
        "operationId": "create_user_terms_and_conditions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          }
        },
        "security": [
          {
            "7qtdtska897dt6kandveusvd88": []
          }
        ]
      }
    },
...
    "/permission-audit-history": {
      "get": {
        "tags": [
          "Permission Audit"
        ],
        "summary": "Get Permission Audit History By User And Application",
        "description": "Retrieve the permission audit history for a given user and application.\n\nArgs:\n    userId (int): The ID of the user for whom the audit history is being requested.\n    applicationId (int): The ID of the application associated with the audit history.\n\nReturns:\n    List[PermissionAuditHistoryRes]: A list of audit history records for the given user and application.",
        "operationId": "get_permission_audit_history_by_user_and_application",
        "security": [
          {
            "7qtdtska897dt6kandveusvd88": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "application_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PermissionAuditHistoryRes"
                  },
                  "title": "Response Get Permission Audit History By User And Application"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/external/v1/users": {
      "get": {
        "tags": [
          "FAM External API"
        ],
        "summary": "Search FAM users",
        "description": "Search FAM users information associated with an application.",
        "operationId": "user_search",
        "security": [
          {
            "7qtdtska897dt6kandveusvd88": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Page number - 1 index",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number - 1 index"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 10
                },
                {
                  "type": "null"
                }
              ],
              "description": "Number of records per page",
              "default": 50,
              "title": "Size"
            },
            "description": "Number of records per page"
          },
          {
            "name": "idpType",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/IDPType"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Identity provider type. Available values: IDIR, BCEID, BCSC",
              "title": "Idptype"
            },
            "description": "Identity provider type. Available values: IDIR, BCEID, BCSC"
          },
          {
            "name": "idpUsername",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "description": "Username from the identity provider",
              "title": "Idpusername"
            },
            "description": "Username from the identity provider"
          },
          {
            "name": "firstName",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "description": "User's first name",
              "title": "Firstname"
            },
            "description": "User's first name"
          },
          {
            "name": "lastName",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 50
                },
                {
                  "type": "null"
                }
              ],
              "description": "User's last name",
              "title": "Lastname"
            },
            "description": "User's last name"
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 5
                },
                {
                  "type": "null"
                }
              ],
              "description": "List of user roles (code, e.g., 'ILCR_SUBMITTER') to filter by (each max 25 chars)",
              "title": "Role"
            },
            "description": "List of user roles (code, e.g., 'ILCR_SUBMITTER') to filter by (each max 25 chars)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtUserSearchPagedResultsSchema_ExtApplicationUserSearchGetSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExtApplicationUserSearchGetSchema": {
        "properties": {
          "firstName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstname"
          },
          "lastName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastname"
          },
          "idpUsername": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Idpusername"
          },
          "idpUserGuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idpuserguid"
          },
          "idpType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IDPType"
              },
              {
                "type": "null"
              }
            ]
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/ExtRoleWithScopeSchema"
            },
            "type": "array",
            "title": "Roles",
            "default": []
          }
        },
        "type": "object",
        "title": "ExtApplicationUserSearchGetSchema",
        "description": "External API response schema.\nFor returning user search results: user detail and roles info."
      },
      "ExtPageResultMetaSchema": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total records counts for query conditions",
            "default": 0
          },
          "pageCount": {
            "type": "integer",
            "title": "Pagecount",
            "description": "Total pages for query records",
            "default": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "title": "Page",
            "description": "Current page number"
          },
          "size": {
            "type": "integer",
            "maximum": 100,
            "minimum": 10,
            "title": "Size",
            "description": "Number of records per page"
          }
        },
        "type": "object",
        "required": [
          "page",
          "size"
        ],
        "title": "ExtPageResultMetaSchema",
        "description": "Response metadata for external API pagination."
      },
...
      "ExtUserSearchPagedResultsSchema_ExtApplicationUserSearchGetSchema_": {
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/ExtPageResultMetaSchema"
          },
          "users": {
            "items": {
              "$ref": "#/components/schemas/ExtApplicationUserSearchGetSchema"
            },
            "type": "array",
            "title": "Users",
            "description": "Paged results"
          }
        },
        "type": "object",
        "required": [
          "meta",
          "users"
        ],
        "title": "ExtUserSearchPagedResultsSchema[ExtApplicationUserSearchGetSchema]"
      },
...
      "FamApplicationUserRoleAssignmentGetSchema": {
        "properties": {
          "user_role_xref_id": {
            "type": "integer",
            "title": "User Role Xref Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "role_id": {
            "type": "integer",
            "title": "Role Id"
          },
          "user": {
            "$ref": "#/components/schemas/FamUserInfoSchema"
          },
          "role": {
            "$ref": "#/components/schemas/FamRoleWithClientSchema"
          },
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "expiry_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          }
        },
        "type": "object",
        "required": [
          "user_role_xref_id",
          "user_id",
          "role_id",
          "user",
          "role",
          "create_date"
        ],
        "title": "FamApplicationUserRoleAssignmentGetSchema"
      },
...
      "FamDistrictSchema": {
        "properties": {
          "org_unit_code": {
            "type": "string",
            "title": "Org Unit Code"
          },
          "orgUnitName": {
            "type": "string",
            "title": "Orgunitname"
          },
          "isExpired": {
            "type": "boolean",
            "title": "Isexpired"
          }
        },
        "type": "object",
        "required": [
          "org_unit_code",
          "orgUnitName",
          "isExpired"
        ],
        "title": "FamDistrictSchema",
        "description": "A BC natural resource district, used to scope roles that have\nrole_type_district set.\n\nField names mirror the org unit source data, which is why orgUnitName and\nisExpired are camelCase while org_unit_code is not."
      },
...
      "FamUserRoleAssignmentCreateRes": {
        "properties": {
          "status_code": {
            "type": "integer",
            "title": "Status Code"
          },
          "detail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FamApplicationUserRoleAssignmentGetSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "email_sending_status": {
            "$ref": "#/components/schemas/EmailSendingStatus",
            "default": "NOT_REQUIRED"
          }
        },
        "type": "object",
        "required": [
          "status_code"
        ],
        "title": "FamUserRoleAssignmentCreateRes"
      },
      "FamUserRoleAssignmentCreateSchema": {
        "properties": {
          "users": {
            "items": {
              "$ref": "#/components/schemas/FamUserRoleAssignmentUserSchema"
            },
            "type": "array",
            "minItems": 1,
            "title": "Users"
          },
          "user_type_code": {
            "$ref": "#/components/schemas/UserType"
          },
          "role_id": {
            "type": "integer",
            "title": "Role Id"
          },
          "forest_client_numbers": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "maxLength": 8,
                  "minLength": 1
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forest Client Numbers"
          },
          "requires_send_user_email": {
            "type": "boolean",
            "title": "Requires Send User Email",
            "default": false
          },
          "expiry_date_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date Date",
            "description": "The expiry date as a string (YYYY-MM-DD), BC timezone. If provided, the role is valid until the end of this day."
          }
        },
        "type": "object",
        "required": [
          "users",
          "user_type_code",
          "role_id"
        ],
        "title": "FamUserRoleAssignmentCreateSchema",
        "description": "Request schema for assigning one or more users to a role, with optional expiry date.\n- users: List of users to assign\n- user_type_code: The type of all users in this request\n- expiry_date_date: The expiry date as a string (YYYY-MM-DD), BC timezone.\n- _expiry_date: Internal use only, timezone-aware datetime, derived from expiry_date_date."
      },
      "FamUserRoleAssignmentRes": {
        "properties": {
          "assignments_detail": {
            "items": {
              "$ref": "#/components/schemas/FamUserRoleAssignmentCreateRes"
            },
            "type": "array",
            "title": "Assignments Detail"
          }
        },
        "type": "object",
        "required": [
          "assignments_detail"
        ],
        "title": "FamUserRoleAssignmentRes"
      },
      "FamUserRoleAssignmentUserSchema": {
        "properties": {
          "user_name": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "User Name"
          },
          "user_guid": {
            "type": "string",
            "maxLength": 32,
            "minLength": 32,
            "title": "User Guid"
          }
        },
        "type": "object",
        "required": [
          "user_name",
          "user_guid"
        ],
        "title": "FamUserRoleAssignmentUserSchema"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IDPType": {
        "type": "string",
        "enum": [
          "IDIR",
          "BCEID",
          "BCSC"
        ],
        "title": "IDPType"
      },
...
      "PermissionAuditHistoryRes": {
        "properties": {
          "create_date": {
            "type": "string",
            "format": "date-time",
            "title": "Create Date"
          },
          "create_user": {
            "type": "string",
            "title": "Create User"
          },
          "change_date": {
            "type": "string",
            "format": "date-time",
            "title": "Change Date"
          },
          "change_performer_user_details": {
            "$ref": "#/components/schemas/PrivilegeChangePerformerSchema"
          },
          "privilege_change_type_code": {
            "$ref": "#/components/schemas/PrivilegeChangeTypeEnum"
          },
          "privilege_details": {
            "$ref": "#/components/schemas/PrivilegeDetailsSchema"
          },
          "privilege_change_audit_id": {
            "type": "integer",
            "title": "Privilege Change Audit Id"
          },
          "change_performer_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Performer User Id"
          },
          "privilege_change_type_description": {
            "type": "string",
            "title": "Privilege Change Type Description"
          }
        },
        "type": "object",
        "required": [
          "create_date",
          "create_user",
          "change_date",
          "change_performer_user_details",
          "privilege_change_type_code",
          "privilege_details",
          "privilege_change_audit_id",
          "change_performer_user_id",
          "privilege_change_type_description"
        ],
        "title": "PermissionAuditHistoryRes",
        "description": "This class is used to transfer data related to the changes made to a user's permissions,\ntypically in the context of an audit trail. It encapsulates details about the change,\nincluding when it occurred, who performed the change, who the change was applied to,\nand the specific details of the permission changes."
      },
      "PrivilegeChangePerformerSchema": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 20,
            "title": "Username"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 250
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "username"
        ],
        "title": "PrivilegeChangePerformerSchema",
        "description": "This schema represents the structure of the `change_user_details` JSON field used in fam_privilege_change_audit.\n\nThe `change_user_details` field captures information about the user who performed a change, including\nthe `username`, `first_name`, `last_name`, and `email`. It is used to record the user details at the time\nof the audit event, ensuring that changes to these details later do not affect the integrity of the audit log.\n\nFor regular users, all fields (`username`, `first_name`, `last_name`, and `email`) are included. However,\nwhen the change is performed by a system account, only the `username` field is present, and it is set to\n\"system\". The schema includes validation logic to enforce this rule.\n\nAttributes:\n    username (str): The username of the user performing the change. For system accounts, this is \"system\".\n    first_name (str, optional): The first name of the user. Not present for system accounts.\n    last_name (str, optional): The last name of the user. Not present for system accounts.\n    email (str, optional): The email address of the user. Not present for system accounts.\n\nValidation:\n    The schema includes a validator to ensure that for system accounts (where `username` is \"system\"),\n    no other fields (`first_name`, `last_name`, `email`) are populated."
      },
      "PrivilegeChangeTypeEnum": {
        "type": "string",
        "enum": [
          "GRANT",
          "REVOKE",
          "UPDATE"
        ],
        "title": "PrivilegeChangeTypeEnum"
      },
...
      "PrivilegeDetailsSchema": {
        "properties": {
          "permission_type": {
            "$ref": "#/components/schemas/PrivilegeDetailsPermissionTypeEnum"
          },
          "roles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PrivilegeDetailsRoleSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "permission_type"
        ],
        "title": "PrivilegeDetailsSchema",
        "description": "This schema represents the structure of the `privilege_details` JSON field used in the `fam_privilege_change_audit` table.\n\nThe `privilege_details` field captures the details of the privileges being changed during a privilege audit event.\nIt includes information about the `permission_type` and, types of permissions, the associated roles and scopes.\n\nAttributes:\n    permission_type (PrivilegeDetailsPermissionTypeEnum): The type of permission being changed.\n    roles (List[PrivilegeDetailsRoleSchema], optional): A list of roles associated with the permission.\n          Required for `END_USER` and `DELEGATED_ADMIN` permission types, and should be omitted for `APPLICATION_ADMIN`.\n\nValidation:\n    The schema includes a validator to ensure that roles are appropriately present or absent based on the `permission_type`."
      },
...
      "UserType": {
        "type": "string",
        "enum": [
          "I",
          "B"
        ],
        "title": "UserType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "7qtdtska897dt6kandveusvd88": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {

            },
            "authorizationUrl": "https://lza-dev-fam-user-pool-domain.auth.ca-central-1.amazoncognito.com/authorize",
            "tokenUrl": "https://lza-dev-fam-user-pool-domain.auth.ca-central-1.amazoncognito.com/token"
          }
        }
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}